+++ packages/desktop/src/code-field/index.ts
@@ -1,0 +1,2 @@
+export * from './CodeFieldProps'
+export * from './CodeField'